Add patch which allows a slot to be removed from the slot-manager
Enable the new support in the gobi plugin for when the pinephone modem
randomly gets restarted/disconnected from the USB bus.

As the slot-manager and sailfish UI need the slot to keep the same name,
in modem.c decrement the modem id when a modem is removed so the name
remains the same.